// src/app/models/zemljiste.ts
import { NamenaZem } from './namena-zem';
import { KatastarskaOpstina } from './katastarska-opstina';

export class Zemljiste {
    id: number;
    povrsina: number;
    knjigVr: number;
    napomena: string;
    katBroj: number;
    katPodbroj: number;
    katDeoParcele: number;
    katPotes: string;
    namenaZem: NamenaZem;
    katastarskaOpstina: KatastarskaOpstina;




}